Moved everything related to userConfig into ConfigManager
