+++ src/main/java/de/t0biii/music/main/Main.java
@@ -1,7 +1,5 @@
 package de.t0biii.music.main;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.logging.Logger;
 import de.t0biii.music.commands.CMD_PlayMusic;
 import de.t0biii.music.domain.ConfigManager;
@@ -13,9 +11,6 @@ import de.t0biii.music.listener.HANDLER_PlayerQuit;
 import de.t0biii.music.listener.HANDLER_SongEndEvent;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.Bukkit;
-import org.bukkit.configuration.InvalidConfigurationException;
-import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
 import de.t0biii.music.domain.bStatsCostum;
@@ -33,7 +28,6 @@ public class Main extends JavaPlugin {
   private int uid = 83541;
   public Updater updater;
   
-  private FileConfiguration userConfigs = null;
 
 
   public ConfigManager cm = new ConfigManager(this);
@@ -55,7 +49,7 @@ public class Main extends JavaPlugin {
     
     cm.loadConfig();
     saveConfig();
-    loadUserConfigsIfNeeded();
+    cm.loadUserConfigsIfNeeded();
     
     Updater();
     Music.createRandomFileDir(this);
@@ -92,42 +86,4 @@ public class Main extends JavaPlugin {
     }
   }
   
-  public void reloadConfigs() {
-	  this.reloadConfig();
-	  cm.loadConfig();
-	  this.saveConfig();
-	  this.loadUserConfigsIfNeeded();
-  }
-  
-  public void loadUserConfigsIfNeeded() {
-	if (this.getConfig().getBoolean("options.allowDisabling")) {
-	  createUserConfigs();
-	}
-  }
-  
-  private void createUserConfigs() {
-	File userConfigsFile = new File(getDataFolder(), "disabledPlayers.yml");
-    if (!userConfigsFile.exists()) {
-      userConfigsFile.getParentFile().mkdirs();
-      saveResource("disabledPlayers.yml", false);
-    }
-	userConfigs = new YamlConfiguration();
-	try {
-        userConfigs.load(userConfigsFile);
-    } catch (IOException | InvalidConfigurationException e) {
-        e.printStackTrace();
-    }
-  }
-  
-  public FileConfiguration getUserConfigs() {
-    return this.userConfigs;
-  }
-  
-  public void saveUserConfigs() {
-	try {
-	  this.userConfigs.save(new File(getDataFolder(), "disabledPlayers.yml"));
-	} catch (IOException e) {
-		e.printStackTrace();
-	}
-  }
 }
